refactor: modularize TurndownService-related code for better maintainability
